pyproject.toml
[project]
name = "babycare"
version = "0.1.0"
description = "A specialized LLM chatbot for baby care knowledge using LangChain and LangGraph with RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.2.1",
    "langchain-community==0.2",
    "langchain_openai>=0.1.7",
    "langchain-pinecone>=0.1.2",
    "langgraph>=0.2.70",
    "langsmith>=0.1.17",
    "pathlib>=1.0.1",
    "pinecone-client>=4.1.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.6.0",
    # Authentication & Database
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pymongo>=4.6.0",
    "motor>=3.3.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "deepeval>=3.6.7",
    "chromadb>=1.0.13",
    "qdrant-client>=1.12.1",
    "tinydb>=4.8.2",
]
